# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/configs": {
            "get": {
                "description": "client가 앱을 실행하기 위해 필요한 구성을 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "구성 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Config"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "gin server의 헬스를 체크합니다.",
                "produces": [
                    "application/json"
                ],
                "summary": "Server Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pong"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "token 클레임에 있는 id 값으로 사용자를 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "사용자 정보 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {AccessToken}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "로그인을 성공 시 JWT token이 발급됩니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "로그인",
                "parameters": [
                    {
                        "description": "로그인 정보",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.JWTToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "404": {
                        "description": ""
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        },
        "/users/login/kakao": {
            "get": {
                "description": "카카오 oauth 인증 시 콜백되는 API 입니다. 환경변수에 주입된 경로로 token을 queryString에 실어서 리디렉션 합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "카카오 로그인",
                "parameters": [
                    {
                        "type": "string",
                        "description": "카카오 로그인 성공 시 전달되는 인가 코드",
                        "name": "code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "",
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "http://localhost:3000/auth-end?accessToken=header.payLoad.signature\u0026refreshToken=header.payLoad.signature"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "description": "token 클레임에 있는 id 값으로 로그아웃 합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "로그아웃",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {AccessToken}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        },
        "/users/signup": {
            "post": {
                "description": "사용자를 생성합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "회원가입",
                "parameters": [
                    {
                        "description": "회원가입을 위한 정보",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserSignUp"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entitys.User"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "/users/1"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        },
        "/users/token/refresh": {
            "post": {
                "description": "엑세스 토큰이 만료되었을때 리프레쉬 토큰을 이용하여 새롭게 발급하기 위하여 사용됩니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "리프레쉬 토큰 발급",
                "parameters": [
                    {
                        "description": "리프레쉬 토큰 정보",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.JWTRefreshToken"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.JWTToken"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        },
        "/users/token/valid": {
            "get": {
                "description": "올바르게 서명된 엑세스 토큰인지 검증합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "엑세스 토큰 검증",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {AccessToken}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entitys.KakaoTalkSocial": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickName": {
                    "type": "string"
                },
                "thumbnailImageUrl": {
                    "description": "이미지의 url을 저장하기 위한 최대 크기로 잡아놓았습니다.",
                    "type": "string"
                }
            }
        },
        "entitys.User": {
            "type": "object",
            "properties": {
                "avatarImage": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "connectedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kakaoTalkSocial": {
                    "$ref": "#/definitions/entitys.KakaoTalkSocial"
                },
                "kakaoTalkSocialsId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.Config": {
            "type": "object",
            "properties": {
                "kakao": {
                    "$ref": "#/definitions/models.Kakao"
                }
            }
        },
        "models.ErrResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "여기에 에러 메시지가 전달됩니다."
                }
            }
        },
        "models.JWTRefreshToken": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string",
                    "example": "header.payLoad.signature"
                }
            }
        },
        "models.JWTToken": {
            "type": "object",
            "required": [
                "accessToken",
                "refreshToken"
            ],
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "header.payLoad.signature"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "header.payLoad.signature"
                }
            }
        },
        "models.Kakao": {
            "type": "object",
            "properties": {
                "redirectUri": {
                    "type": "string",
                    "example": "http://localhost:8080/api/users/login/kakao"
                },
                "restApiKey": {
                    "type": "string",
                    "example": "111aaaa1a111aaa1a111a1a1a11a1a11"
                }
            }
        },
        "models.Pong": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "pong"
                }
            }
        },
        "models.UserInfo": {
            "type": "object",
            "required": [
                "avatarImage",
                "email",
                "id",
                "name"
            ],
            "properties": {
                "avatarImage": {
                    "type": "string",
                    "example": "base64으로 인코딩된 이미지"
                },
                "email": {
                    "type": "string",
                    "example": "user01@test.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "사용자01"
                }
            }
        },
        "models.UserLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user01@test.com"
                },
                "password": {
                    "type": "string",
                    "example": "test1!"
                }
            }
        },
        "models.UserSignUp": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "passwordConfirm"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user01@test.com"
                },
                "name": {
                    "type": "string",
                    "example": "사용자01"
                },
                "password": {
                    "type": "string",
                    "example": "test1!"
                },
                "passwordConfirm": {
                    "type": "string",
                    "example": "test1!"
                }
            }
        }
    }
}